+++ tests/IO/EwkbWriterTest.php
@@ -52,22 +52,50 @@ class EwkbWriterTest extends EwkbAbstractTestCase
         }
     }
 
-    #[DataProvider('providerWriteEmptyPointThrowsException')]
-    public function testWriteEmptyPointThrowsException(Point $point) : void
+    #[DataProvider('providerWriteEmptyPointWithoutNanSupportThrowsException')]
+    public function testWriteEmptyPointWithoutNanSupportThrowsException(Point $point) : void
     {
-        $writer = new EwkbWriter();
+        $writer = new EwkbWriter(supportEmptyPointWithNan: false);
 
         $this->expectException(GeometryIoException::class);
+        $this->expectExceptionMessage(
+            'Empty points have no WKB representation. If you want to output empty points with NaN coordinates ' .
+            '(PostGIS-style), enable the $supportEmptyPointWithNan option.',
+        );
+
         $writer->write($point);
     }
 
-    public static function providerWriteEmptyPointThrowsException() : array
+    public static function providerWriteEmptyPointWithoutNanSupportThrowsException() : array
     {
         return [
             [Point::xyEmpty()],
             [Point::xyzEmpty()],
             [Point::xymEmpty()],
             [Point::xyzmEmpty()]
+        ];
+    }
+
+    #[DataProvider('providerWriteEmptyPointWithNanSupport')]
+    public function testWriteEmptyPointWithNanSupport(Point $point, WkbByteOrder $byteOrder, string $expectedHex) : void
+    {
+        $writer = new EwkbWriter(byteOrder: $byteOrder);
+
+        $actualHex = bin2hex($writer->write($point));
+        self::assertSame($expectedHex, $actualHex);
+    }
+
+    public static function providerWriteEmptyPointWithNanSupport() : array
+    {
+        return [
+            [Point::xyEmpty(), WkbByteOrder::BigEndian, '00000000017ff80000000000007ff8000000000000'],
+            [Point::xyEmpty(), WkbByteOrder::LittleEndian, '0101000000000000000000f87f000000000000f87f'],
+            [Point::xyzEmpty(), WkbByteOrder::BigEndian, '00800000017ff80000000000007ff80000000000007ff8000000000000'],
+            [Point::xyzEmpty(), WkbByteOrder::LittleEndian, '0101000080000000000000f87f000000000000f87f000000000000f87f'],
+            [Point::xymEmpty(), WkbByteOrder::BigEndian, '00400000017ff80000000000007ff80000000000007ff8000000000000'],
+            [Point::xymEmpty(), WkbByteOrder::LittleEndian, '0101000040000000000000f87f000000000000f87f000000000000f87f'],
+            [Point::xyzmEmpty(), WkbByteOrder::BigEndian, '00c00000017ff80000000000007ff80000000000007ff80000000000007ff8000000000000'],
+            [Point::xyzmEmpty(), WkbByteOrder::LittleEndian, '01010000c0000000000000f87f000000000000f87f000000000000f87f000000000000f87f'],
         ];
     }
 }